Force UTC for is_due comparisons.
In order to remove any ambiguity with DST time comparison force all
time comparisons to be in UTC.  One side of this is_due comparison is
last_run_at which is written and read to from the database.  This field
is stored in naive UTC.  The other side of the comparison is the
current time.  Because the comparison actually happens in celery we
pass in a function that returns UTC when it wants to get the current
time for comparison to last_run_at

Normally we might store last_run_at in the database using django's
timezone but this can cause an ambiguity during the DST transition.  I
also investigated storing UTC in the database and then localizing the
time to django's timezone on the way out.  This appears to work, because
a comparison of a pre-dst to post-dst values will have different utc
offsets and thus have the correct time delta.  But, a peculiarity in
celery's remaining() function negates this.